+++ packages/playground/babel.config.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+    presets: [
+      ['@babel/preset-env', { targets: { node: 'current' } }],
+      '@babel/preset-typescript',
+    ],
+  };
+
