+++ packages/playground/src/StudyConfiguration/custom/simjs/lib/model.js
