+++ packages/playground/src/StudyConfiguration/custom/simjs/lib/sim.d.ts
