+++ packages/playground/src/StudyConfiguration/custom/simjs/lib/sim.js
