+++ packages/playground/src/StudyConfiguration/custom/simjs/lib/stats.js
