+++ packages/playground/src/StudyConfiguration/custom/simjs/sim.js
