+++ packages/playground/src/StudyConfiguration/custom/timeline/Sim.d.ts
