+++ packages/playground/src/StudyConfiguration/custom/timeline/Timeline.ts
@@ -1,5 +1,6 @@
 import { RtBoolean, RtObject } from '@freon4dsl/core';
 import { ScheduledEvent } from './ScheduledEvent';
+import { Event } from '../../language/gen/index';
 
 /*
  * A Timeline records the events and the days they occur on.
@@ -48,9 +49,9 @@ export class Timeline extends RtObject{
     return timelineDay ? timelineDay.events : [];
   }
 
-  getLastInstanceForThisEvent(scheduledEvent: ScheduledEvent): EventInstance {
+  getLastInstanceForThisEvent(event: Event): EventInstance {
     let allEventInstances = this.days.flatMap(day => day.events);
-    let eventInstances = allEventInstances.filter(event => event.scheduledEvent.name === scheduledEvent.name);
+    let eventInstances = allEventInstances.filter(event => event.scheduledEvent.name() === event.name);
     return eventInstances[eventInstances.length - 1]; // TODO: sort by day and get the most recent
   }
