+++ packages/playground/src/StudyConfiguration/custom/timeline/visjs-timeline.html
