+++ packages/playground/src/StudyConfiguration/custom/types/simjs.d.ts
