+++ packages/playground/src/StudyConfiguration/custom/utils/SimpleLogger.js
